+++ lib/src/run_count_result.dart
@@ -1,0 +1,9 @@
+class RunCountResult {
+  final int childrenPerRun;
+  final bool? additionalChildOnExistingRun;
+
+  const RunCountResult({
+    required this.childrenPerRun,
+    this.additionalChildOnExistingRun,
+  });
+}