+++ test/flutter_extended_wrap_test.dart
@@ -4,8 +4,11 @@ import 'package:flutter_extended_wrap/flutter_extended_wrap.dart';
 
 void main() {
   test('test max children count per run', () {
-    const ExtendedWrap extendedWrap =
-        ExtendedWrap(childSizeInDirection: 50, spacing: 10, minFreeSpacePerRun: 100, children: <Widget>[]);
+    const ExtendedWrap extendedWrap = ExtendedWrap(
+        childSizeInDirection: 50,
+        spacing: 10,
+        minFreeSpacePerRun: 100,
+        children: <Widget>[]);
     int manyMaxChildren = extendedWrap.getMaxChildrenCountPerRun(500);
     expect(manyMaxChildren, 6, reason: "should be 6");
     manyMaxChildren = extendedWrap.getMaxChildrenCountPerRun(1378);
@@ -14,7 +17,8 @@ void main() {
     expect(fewMaxChildren, 1, reason: "should be 1");
     fewMaxChildren = extendedWrap.getMaxChildrenCountPerRun(150);
     expect(fewMaxChildren, 1, reason: "should still be 1");
-    expect(() => extendedWrap.getMaxChildrenCountPerRun(140), throwsA((Object e) => e is AssertionError));
+    expect(() => extendedWrap.getMaxChildrenCountPerRun(140),
+        throwsA((Object e) => e is AssertionError));
   });
   // todo: many more tests could follow here in the future...
 }