+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.0.3] - 2023-11-28
+
+* widened dart constraints
+* added new alignment spaceCenterAndAround and renamed spaceCenterOrAround to spaceCenterOrAlignCenter
+
 ## [0.0.2] - 2023-11-27
 
 * updated README.md